Update the src submodule's SHA to 2b19082
In T146389 the pivot src repository was updated with the build
directory, containing type script files compiled to javascript.
The submodule SHA was not updated in the pivot/deploy repository
so git submodule update --init is not able to pick up the latest
version of the module before deploying.

Related change: https://gerrit.wikimedia.org/r/#/c/315067/

Bug: T146389
Change-Id: I96baa9da2a05568eb1bddb8a61ea0a3866c64cf1